refactor(infra): let EF bind Money through its existing constructor
- drop the parameterless constructor added to Money; EF Core matches the
  private constructor by parameter name, so the domain type is untouched
- rename the schema test to describe what it actually asserts
- compare column precision without casting

+++ tests/Envanex.IntegrationTests/Persistence/SchemaPrecisionTests.cs
@@ -29,7 +29,7 @@ public sealed class SchemaPrecisionTests : IAsyncLifetime
                 """)
             .SingleAsync();
 
-        ((int)result.Precision).ShouldBe(18);
+        result.Precision.ShouldBe((byte)18);
         result.Scale.ShouldBe(4);
     }
 
@@ -48,7 +48,7 @@ public sealed class SchemaPrecisionTests : IAsyncLifetime
                 """)
             .SingleAsync();
 
-        ((int)result.Precision).ShouldBe(18);
+        result.Precision.ShouldBe((byte)18);
         result.Scale.ShouldBe(6);
     }
